+++ src/main/java/net/jodah/failsafe/Callables.java
@@ -4,9 +4,6 @@ import java.util.concurrent.Callable;
 import java.util.concurrent.Semaphore;
 import java.util.function.BiConsumer;
 
-import net.jodah.failsafe.event.ContextualResultListener;
-import net.jodah.failsafe.event.FailureListener;
-import net.jodah.failsafe.event.ResultListener;
 import net.jodah.failsafe.function.AsyncCallable;
 import net.jodah.failsafe.function.AsyncRunnable;
 import net.jodah.failsafe.function.CheckedRunnable;
@@ -34,37 +31,6 @@ final class Callables {
     void inject(AsyncExecution execution) {
       this.execution = execution;
     }
-  }
-
-  static <T> Callable<T> of(final ContextualResultListener<T, Throwable> listener, final T result,
-      final Throwable failure, final ExecutionContext context) {
-    return new Callable<T>() {
-      @Override
-      public T call() {
-        listener.onResult(result, failure, context);
-        return null;
-      }
-    };
-  }
-
-  static <T> Callable<T> of(final FailureListener<Throwable> listener, final Throwable failure) {
-    return new Callable<T>() {
-      @Override
-      public T call() {
-        listener.onFailure(failure);
-        return null;
-      }
-    };
-  }
-
-  static <T> Callable<T> of(final ResultListener<T, Throwable> listener, final T result, final Throwable failure) {
-    return new Callable<T>() {
-      @Override
-      public T call() {
-        listener.onResult(result, failure);
-        return null;
-      }
-    };
   }
 
   static <T> Callable<T> of(final CheckedRunnable runnable) {
@@ -291,4 +257,5 @@ final class Callables {
       }
     };
   }
+
 }